requirements.txt
datetime
paddlepaddle
imgaug
paddleocr
pillow
